refactor(swapExample): factor out flows
