+++ packages/orchestration/src/examples/swapExample.contract.js
@@ -1,12 +1,10 @@
 import { StorageNodeShape } from '@agoric/internal';
 import { TimerServiceShape } from '@agoric/time';
 import { M } from '@endo/patterns';
-import { orcUtils } from '../utils/orc.js';
 import { withOrchestration } from '../utils/start-helper.js';
+import * as flows from './swapExample.flows.js';
 
 /**
- * @import {LocalTransfer} from '../utils/zoe-tools.js';
- * @import {Orchestrator, CosmosValidatorAddress, OrchestrationFlow} from '../types.js'
  * @import {TimerService} from '@agoric/time';
  * @import {LocalChain} from '@agoric/vats/src/localchain.js';
  * @import {Remote} from '@agoric/internal';
@@ -15,50 +13,6 @@ import { withOrchestration } from '../utils/start-helper.js';
  * @import {Zone} from '@agoric/zone';
  * @import {OrchestrationTools} from '../utils/start-helper.js';
  */
-
-/**
- * @satisfies {OrchestrationFlow}
- * @param {Orchestrator} orch
- * @param {object} ctx
- * @param {LocalTransfer} ctx.localTransfer
- * @param {ZCFSeat} seat
- * @param {object} offerArgs
- * @param {Amount<'nat'>} offerArgs.staked
- * @param {CosmosValidatorAddress} offerArgs.validator
- */
-const stakeAndSwapFn = async (orch, { localTransfer }, seat, offerArgs) => {
-  const { give } = seat.getProposal();
-
-  const omni = await orch.getChain('omniflixhub');
-  const agoric = await orch.getChain('agoric');
-
-  const [omniAccount, localAccount] = await Promise.all([
-    omni.makeAccount(),
-    agoric.makeAccount(),
-  ]);
-
-  const omniAddress = omniAccount.getAddress();
-
-  // deposit funds from user seat to LocalChainAccount
-  await localTransfer(seat, localAccount, give);
-  seat.exit();
-
-  // build swap instructions with orcUtils library
-  const transferMsg = orcUtils.makeOsmosisSwap({
-    destChain: 'omniflixhub',
-    destAddress: omniAddress,
-    amountIn: give.Stable,
-    brandOut: /** @type {any} */ ('FIXME'),
-    slippage: 0.03,
-  });
-
-  try {
-    await localAccount.transferSteps(give.Stable, transferMsg);
-    await omniAccount.delegate(offerArgs.validator, offerArgs.staked);
-  } catch (e) {
-    console.error(e);
-  }
-};
 
 /** @type {ContractMeta<typeof start>} */
 export const meta = {
@@ -98,20 +52,23 @@ export const makeNatAmountShape = (brand, min) =>
  * @param {Zone} zone
  * @param {OrchestrationTools} tools
  */
-const contract = async (zcf, privateArgs, zone, { orchestrate, zoeTools }) => {
+const contract = async (
+  zcf,
+  privateArgs,
+  zone,
+  { zoeTools, orchestrateAll },
+) => {
   const { brands } = zcf.getTerms();
 
-  /** deprecated historical example */
-  const swapAndStakeHandler = orchestrate(
-    'LSTTia',
-    { zcf, localTransfer: zoeTools.localTransfer },
-    stakeAndSwapFn,
-  );
+  const handlers = orchestrateAll(flows, {
+    zcf,
+    localTransfer: zoeTools.localTransfer,
+  });
 
   const publicFacet = zone.exo('publicFacet', undefined, {
     makeSwapAndStakeInvitation() {
       return zcf.makeInvitation(
-        swapAndStakeHandler,
+        handlers.swapAndStake,
         'Swap for TIA and stake',
         undefined,
         harden({
